+++ core/src/main/java/me/mykindos/betterpvp/core/item/model/ArmorItem.java
@@ -20,7 +20,7 @@ public abstract class ArmorItem extends BaseItem implements Reloadable {
     protected ArmorItem(BPvPPlugin plugin, String name, ItemStack model, ItemRarity rarity) {
         super(name, model, ItemGroup.ARMOR, rarity);
         this.plugin = plugin;
-        addSerializableComponent(new StatContainerComponent().withBaseStat(new ItemStat<>(StatTypes.HEALTH, 1)));
+        addSerializableComponent(new StatContainerComponent().withBaseStat(new ItemStat<>(StatTypes.HEALTH, 1d)));
         addSerializableComponent(new DurabilityComponent(500));
         addSerializableComponent(new RuneContainerComponent(0, 0));
     }
@@ -28,9 +28,9 @@ public abstract class ArmorItem extends BaseItem implements Reloadable {
     @Override
     public void reload() {
         final Config config = Config.item(plugin, this);
-        final int health = config.getConfig("health.base", 1, Integer.class);
-        final int healthMin = config.getConfig("health.min", 0, Integer.class);
-        final int healthMax = config.getConfig("health.max", 2, Integer.class);
+        final double health = config.getConfig("health.base", 1, Double.class);
+        final double healthMin = config.getConfig("health.min", 0, Double.class);
+        final double healthMax = config.getConfig("health.max", 2, Double.class);
         getComponent(StatContainerComponent.class).ifPresent(statContainer -> {
             statContainer.getStat(StatTypes.HEALTH).ifPresent(healthStat -> {
                 statContainer.withBaseStat(healthStat.withValue(health).withRanges(healthMin, healthMax));
